requirements.txt
 psycopg2-binary
 pygsheets
 python-decouple
 python-pagarme
 pytz
 sqlalchemy